+++ VERSION
@@ -1,1 +1,1 @@
-1.12.2-next.3
+1.12.2-next.4
