+++ examples/flat-config-react-dom-ts/package.json
@@ -23,7 +23,7 @@
     "eslint": "^9.9.1",
     "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-react-refresh": "^0.4.11",
-    "typescript": "^5.5.4",
+    "typescript": "^5.7.0-dev.20240826",
     "typescript-eslint": "^8.2.0",
     "vite": "^5.4.2"
   }
