+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@eslint-react/monorepo",
-  "version": "1.12.2-next.3",
+  "version": "1.12.2-next.4",
   "private": true,
   "description": "A series of composable ESLint plugins for libraries and frameworks that use React as a UI runtime.",
   "keywords": [
@@ -102,7 +102,7 @@
     "typedoc": "^0.26.6",
     "typedoc-plugin-markdown": "^4.2.6",
     "typedoc-plugin-rename-defaults": "^0.7.1",
-    "typescript": "^5.5.4",
+    "typescript": "^5.7.0-dev.20240826",
     "typescript-eslint": "^8.2.0",
     "vitest": "^2.0.5",
     "yaml-eslint-parser": "^1.2.3"
@@ -119,7 +119,7 @@
       "is-core-module": "npm:@nolyfill/is-core-module@^1",
       "ts-api-utils": "^1.3.0",
       "typedarray": "npm:@nolyfill/typedarray@^1",
-      "typescript": "^5.5.4"
+      "typescript": "^5.7.0-dev.20240826"
     },
     "overridesMeta": {
       "esbuild": "override to single version",
