+++ packages/core/docs/functions/isThisSetState.md
@@ -1,0 +1,17 @@
+[**@eslint-react/core**](../README.md) • **Docs**
+
+***
+
+[@eslint-react/core](../README.md) / isThisSetState
+
+# Function: isThisSetState()
+
+> **isThisSetState**(`node`): `boolean`
+
+## Parameters
+
+• **node**: `CallExpression`
+
+## Returns
+
+`boolean`
