+++ packages/core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@eslint-react/core",
-  "version": "1.12.2-next.3",
+  "version": "1.12.2-next.4",
   "description": "ESLint React's ESLint utility module for static analysis of React core APIs and Patterns.",
   "homepage": "https://github.com/rel1cx/eslint-react",
   "bugs": {
