+++ packages/core/src/state/index.ts
@@ -1,1 +1,2 @@
+export * from "./set-state";
 export * from "./state-kind";
