+++ packages/core/src/state/set-state.ts
@@ -1,0 +1,13 @@
+import { isThisExpression } from "@eslint-react/ast";
+import type { TSESTree } from "@typescript-eslint/types";
+import { AST_NODE_TYPES } from "@typescript-eslint/types";
+
+export function isThisSetState(node: TSESTree.CallExpression) {
+  const { callee } = node;
+  return (
+    callee.type === AST_NODE_TYPES.MemberExpression
+    && isThisExpression(callee.object)
+    && callee.property.type === AST_NODE_TYPES.Identifier
+    && callee.property.name === "setState"
+  );
+}
