+++ packages/plugins/eslint-plugin-react-hooks-extra/package.json
@@ -1,6 +1,6 @@
 {
   "name": "eslint-plugin-react-hooks-extra",
-  "version": "1.12.2-next.3",
+  "version": "1.12.2-next.4",
   "description": "ESLint React's ESLint plugin for React Hooks related rules.",
   "keywords": [
     "eslint",
