+++ packages/plugins/eslint-plugin-react-web-api/package.json
@@ -1,6 +1,6 @@
 {
   "name": "eslint-plugin-react-web-api",
-  "version": "1.12.2-next.3",
+  "version": "1.12.2-next.4",
   "description": "ESLint React's ESLint plugin for interacting with Web APIs",
   "keywords": [
     "eslint",
