+++ packages/plugins/eslint-plugin-react-x/package.json
@@ -1,6 +1,6 @@
 {
   "name": "eslint-plugin-react-x",
-  "version": "1.12.2-next.3",
+  "version": "1.12.2-next.4",
   "description": "A set of composable ESLint rules for libraries and frameworks that use React as a UI runtime.",
   "keywords": [
     "eslint",
