+++ packages/plugins/eslint-plugin/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@eslint-react/eslint-plugin",
-  "version": "1.12.2-next.3",
+  "version": "1.12.2-next.4",
   "description": "The main ESLint plugin of ESLint React. Contains all the rules and presets of ESLint React.",
   "keywords": [
     "eslint",
