+++ packages/tools/docs/namespaces/O/functions/productMany.md
@@ -16,7 +16,7 @@
 
 • **self**: [`Option`](../type-aliases/Option.md)\<`A`\>
 
-• **collection**: `Iterable`\<[`Option`](../type-aliases/Option.md)\<`A`\>\>
+• **collection**: `Iterable`\<[`Option`](../type-aliases/Option.md)\<`A`\>, `any`, `any`\>
 
 ## Returns
