+++ packages/types/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@eslint-react/types",
-  "version": "1.12.2-next.3",
+  "version": "1.12.2-next.4",
   "description": "ESLint React's type definitions.",
   "homepage": "https://github.com/rel1cx/eslint-react",
   "bugs": {
