+++ packages/utilities/var/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@eslint-react/var",
-  "version": "1.12.2-next.3",
+  "version": "1.12.2-next.4",
   "description": "TSESTree AST utility module for static analysis of variables",
   "homepage": "https://github.com/rel1cx/eslint-react",
   "bugs": {
