+++ website/package.json
@@ -42,7 +42,7 @@
     "remark-frontmatter": "^5.0.0",
     "remark-gfm": "^4.0.0",
     "sass": "^1.77.8",
-    "typescript": "^5.5.4",
+    "typescript": "^5.7.0-dev.20240826",
     "typescript-eslint": "^8.2.0",
     "vite": "^5.4.2"
   }
